+++ tests/test_queryset.py
@@ -140,6 +140,16 @@ class TestQueryset(test.TestCase):
             await IntFields.all().order_by("intnum").filter(intnum__gte=400).first(), None
         )
 
+    async def test_get_or_none(self):
+        # Test first
+        self.assertEqual(
+            (await IntFields.all().order_by("intnum").get_or_none(intnum__gte=40)).intnum, 40
+        )
+
+        self.assertEqual(
+            await IntFields.all().order_by("intnum").get_or_none(intnum__gte=400), None
+        )
+
     async def test_get(self):
         await IntFields.filter(intnum__gte=70).update(intnum_null=80)