+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Unreleased
 - ``QuerySet.Update()`` now returns the count of the no of rows affected. Note, that
 - ``QuerySet.Delete()`` now returns the count of the no of rows deleted.
 - Note that internal API of ``db_connection.execute_query()`` now returns ``rows_affected, results``. (This is informational only)
+- Added ``get_or_none(...)`` as syntactic sugar for ``filter(...).first()``
 
 
 0.15.6